Fix token usage tests without Electron runtime
